refactor: isolate campaign lifecycle validation
